Agent Legacy class
