+++ packages/core/src/agent/agent.ts
@@ -2,13 +2,12 @@ import { randomUUID } from 'crypto';
 import type { WritableStream } from 'stream/web';
 import type { CoreMessage, TextPart, UIMessage } from '@internal/ai-sdk-v4/message';
 import type { StreamObjectResult } from '@internal/ai-sdk-v4/model';
-import type { Tool } from '@internal/ai-sdk-v4/tool';
 import deepEqual from 'fast-deep-equal';
 import type { JSONSchema7 } from 'json-schema';
 import { z } from 'zod';
 import type { ZodSchema } from 'zod';
 import type { MastraPrimitives, MastraUnion } from '../action';
-import { AISpanType, getOrCreateSpan, getValidTraceId } from '../ai-tracing';
+import { AISpanType, getOrCreateSpan } from '../ai-tracing';
 import type { AISpan, TracingContext, TracingOptions, TracingProperties } from '../ai-tracing';
 import { MastraBase } from '../base';
 import { MastraError, ErrorDomain, ErrorCategory } from '../error';
@@ -22,27 +21,9 @@ import type {
 import { runScorer } from '../evals/hooks';
 import { resolveModelConfig } from '../llm';
 import { MastraLLMV1 } from '../llm/model';
-import type {
-  GenerateObjectWithMessagesArgs,
-  GenerateTextWithMessagesArgs,
-  GenerateReturn,
-  GenerateObjectResult,
-  GenerateTextResult,
-  StreamTextWithMessagesArgs,
-  StreamObjectWithMessagesArgs,
-  StreamReturn,
-  ToolSet,
-  OriginalStreamTextOnFinishEventArg,
-  OriginalStreamObjectOnFinishEventArg,
-  StreamTextResult,
-} from '../llm/model/base.types';
+import type { GenerateObjectResult, GenerateTextResult, StreamTextResult } from '../llm/model/base.types';
 import { MastraLLMVNext } from '../llm/model/model.loop';
-import type {
-  TripwireProperties,
-  MastraLanguageModel,
-  MastraLanguageModelV2,
-  MastraModelConfig,
-} from '../llm/model/shared.types';
+import type { MastraLanguageModel, MastraLanguageModelV2, MastraModelConfig } from '../llm/model/shared.types';
 import { RegisteredLogger } from '../logger';
 import { networkLoop } from '../loop/network';
 import type { Mastra } from '../mastra';
@@ -63,6 +44,7 @@ import type { ToolOptions } from '../utils';
 import type { CompositeVoice } from '../voice';
 import { DefaultVoice } from '../voice';
 import type { Workflow } from '../workflows';
+import { AgentLegacyHandler } from './agent-legacy';
 import type { AgentExecutionOptions, InnerAgentExecutionOptions, MultiPrimitiveExecutionOptions } from './agent.types';
 import { MessageList } from './message-list';
 import type { MessageInput, MessageListInput, UIMessageWithMetadata, MastraDBMessage } from './message-list';
@@ -74,13 +56,13 @@ import type {
   AgentStreamOptions,
   ToolsetsInput,
   ToolsInput,
-  AgentMemoryOption,
   AgentModelManagerConfig,
   AgentCreateOptions,
   AgentExecuteOnFinishOptions,
   AgentInstructions,
   DynamicAgentInstructions,
 } from './types';
+import { resolveThreadIdFromArgs } from './utils';
 import { createPrepareStreamWorkflow } from './workflows/prepare-stream';
 
 export type MastraLLM = MastraLLMV1 | MastraLLMVNext;
@@ -98,19 +80,6 @@ function resolveMaybePromise<T, R = void>(value: T | Promise<T> | PromiseLike<T>
   }
 
   return cb(value as T);
-}
-
-// Helper to resolve threadId from args (supports both new and old API)
-function resolveThreadIdFromArgs(args: {
-  memory?: AgentMemoryOption;
-  threadId?: string;
-}): (Partial<StorageThreadType> & { id: string }) | undefined {
-  if (args?.memory?.thread) {
-    if (typeof args.memory.thread === 'string') return { id: args.memory.thread };
-    if (typeof args.memory.thread === 'object' && args.memory.thread.id) return args.memory.thread;
-  }
-  if (args?.threadId) return { id: args.threadId };
-  return undefined;
 }
 
 /**
@@ -155,6 +124,7 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
   #inputProcessors?: DynamicArgument<InputProcessor[]>;
   #outputProcessors?: DynamicArgument<OutputProcessor[]>;
   readonly #options?: AgentCreateOptions;
+  #legacyHandler?: AgentLegacyHandler;
 
   // This flag is for agent network messages. We should change the agent network formatting and remove this flag after.
   private _agentNetworkAppend = false;
@@ -654,6 +624,44 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
    */
   public getDescription(): string {
     return this.#description ?? '';
+  }
+
+  /**
+   * Gets the legacy handler instance, initializing it lazily if needed.
+   * @internal
+   */
+  private getLegacyHandler(): AgentLegacyHandler {
+    if (!this.#legacyHandler) {
+      this.#legacyHandler = new AgentLegacyHandler({
+        logger: this.logger,
+        name: this.name,
+        id: this.id,
+        mastra: this.#mastra,
+        getDefaultGenerateOptionsLegacy: this.getDefaultGenerateOptionsLegacy.bind(this),
+        getDefaultStreamOptionsLegacy: this.getDefaultStreamOptionsLegacy.bind(this),
+        hasOwnMemory: this.hasOwnMemory.bind(this),
+        getInstructions: async (options: { requestContext: RequestContext }) => {
+          const result = await this.getInstructions(options);
+          return result;
+        },
+        getLLM: this.getLLM.bind(this) as any,
+        getMemory: this.getMemory.bind(this),
+        convertTools: this.convertTools.bind(this),
+        getMemoryMessages: this.getMemoryMessages.bind(this),
+        __runInputProcessors: this.__runInputProcessors.bind(this),
+        getMostRecentUserMessage: this.getMostRecentUserMessage.bind(this),
+        genTitle: this.genTitle.bind(this),
+        resolveTitleGenerationConfig: this.resolveTitleGenerationConfig.bind(this),
+        saveStepMessages: this.saveStepMessages.bind(this),
+        convertInstructionsToString: this.#convertInstructionsToString.bind(this),
+        tracingPolicy: this.#options?.tracingPolicy,
+        _agentNetworkAppend: this._agentNetworkAppend,
+        listResolvedOutputProcessors: this.listResolvedOutputProcessors.bind(this),
+        __runOutputProcessors: this.__runOutputProcessors.bind(this),
+        runScorers: this.#runScorers.bind(this),
+      });
+    }
+    return this.#legacyHandler;
   }
 
   /**
@@ -2048,541 +2056,6 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
     }
   }
 
-  /**
-   * Prepares message list and tools before LLM execution and handles memory persistence after.
-   * @internal
-   */
-  __primitive({
-    instructions,
-    messages,
-    context,
-    thread,
-    memoryConfig,
-    resourceId,
-    runId,
-    toolsets,
-    clientTools,
-    requestContext,
-    saveQueueManager,
-    writableStream,
-    methodType,
-    tracingContext,
-    tracingOptions,
-  }: {
-    instructions: AgentInstructions;
-    toolsets?: ToolsetsInput;
-    clientTools?: ToolsInput;
-    resourceId?: string;
-    thread?: (Partial<StorageThreadType> & { id: string }) | undefined;
-    memoryConfig?: MemoryConfig;
-    context?: CoreMessage[];
-    runId?: string;
-    messages: MessageListInput;
-    requestContext: RequestContext;
-    saveQueueManager: SaveQueueManager;
-    writableStream?: WritableStream<ChunkType>;
-    methodType: 'generate' | 'stream';
-    tracingContext?: TracingContext;
-    tracingOptions?: TracingOptions;
-  }) {
-    return {
-      before: async () => {
-        if (process.env.NODE_ENV !== 'test') {
-          this.logger.debug(`[Agents:${this.name}] - Starting generation`, { runId });
-        }
-
-        const agentAISpan = getOrCreateSpan({
-          type: AISpanType.AGENT_RUN,
-          name: `agent run: '${this.id}'`,
-          input: {
-            messages,
-          },
-          attributes: {
-            agentId: this.id,
-            instructions: this.#convertInstructionsToString(instructions),
-            availableTools: [
-              ...(toolsets ? Object.keys(toolsets) : []),
-              ...(clientTools ? Object.keys(clientTools) : []),
-            ],
-          },
-          metadata: {
-            runId,
-            resourceId,
-            threadId: thread ? thread.id : undefined,
-          },
-          tracingPolicy: this.#options?.tracingPolicy,
-          tracingOptions,
-          tracingContext,
-          requestContext,
-        });
-
-        const innerTracingContext: TracingContext = { currentSpan: agentAISpan };
-
-        const memory = await this.getMemory({ requestContext });
-
-        const toolEnhancements = [
-          // toolsets
-          toolsets && Object.keys(toolsets || {}).length > 0
-            ? `toolsets present (${Object.keys(toolsets || {}).length} tools)`
-            : undefined,
-
-          // memory tools
-          memory && resourceId ? 'memory and resourceId available' : undefined,
-        ]
-          .filter(Boolean)
-          .join(', ');
-        this.logger.debug(`[Agent:${this.name}] - Enhancing tools: ${toolEnhancements}`, {
-          runId,
-          toolsets: toolsets ? Object.keys(toolsets) : undefined,
-          clientTools: clientTools ? Object.keys(clientTools) : undefined,
-          hasMemory: !!memory,
-          hasResourceId: !!resourceId,
-        });
-
-        const threadId = thread?.id;
-
-        const convertedTools = await this.convertTools({
-          toolsets,
-          clientTools,
-          threadId,
-          resourceId,
-          runId,
-          requestContext,
-          tracingContext: innerTracingContext,
-          writableStream,
-          methodType,
-        });
-
-        const messageList = new MessageList({
-          threadId,
-          resourceId,
-          generateMessageId: this.#mastra?.generateId?.bind(this.#mastra),
-          // @ts-ignore Flag for agent network messages
-          _agentNetworkAppend: this._agentNetworkAppend,
-        })
-          .addSystem(instructions || (await this.getInstructions({ requestContext })))
-          .add(context || [], 'context');
-
-        if (!memory || (!threadId && !resourceId)) {
-          messageList.add(messages, 'user');
-          const { tripwireTriggered, tripwireReason } = await this.__runInputProcessors({
-            requestContext,
-            tracingContext: innerTracingContext,
-            messageList,
-          });
-          return {
-            messageObjects: tripwireTriggered ? [] : messageList.get.all.prompt(),
-            convertedTools,
-            threadExists: false,
-            thread: undefined,
-            messageList,
-            agentAISpan,
-            ...(tripwireTriggered && {
-              tripwire: true,
-              tripwireReason,
-            }),
-          };
-        }
-        if (!threadId || !resourceId) {
-          const mastraError = new MastraError({
-            id: 'AGENT_MEMORY_MISSING_RESOURCE_ID',
-            domain: ErrorDomain.AGENT,
-            category: ErrorCategory.USER,
-            details: {
-              agentName: this.name,
-              threadId: threadId || '',
-              resourceId: resourceId || '',
-            },
-            text: `A resourceId and a threadId must be provided when using Memory. Saw threadId "${threadId}" and resourceId "${resourceId}"`,
-          });
-          this.logger.trackException(mastraError);
-          this.logger.error(mastraError.toString());
-          agentAISpan?.error({ error: mastraError });
-          throw mastraError;
-        }
-        const store = memory.constructor.name;
-        this.logger.debug(
-          `[Agent:${this.name}] - Memory persistence enabled: store=${store}, resourceId=${resourceId}`,
-          {
-            runId,
-            resourceId,
-            threadId,
-            memoryStore: store,
-          },
-        );
-
-        let threadObject: StorageThreadType | undefined = undefined;
-        const existingThread = await memory.getThreadById({ threadId });
-        if (existingThread) {
-          if (
-            (!existingThread.metadata && thread.metadata) ||
-            (thread.metadata && !deepEqual(existingThread.metadata, thread.metadata))
-          ) {
-            threadObject = await memory.saveThread({
-              thread: { ...existingThread, metadata: thread.metadata },
-              memoryConfig,
-            });
-          } else {
-            threadObject = existingThread;
-          }
-        } else {
-          threadObject = await memory.createThread({
-            threadId,
-            metadata: thread.metadata,
-            title: thread.title,
-            memoryConfig,
-            resourceId,
-            saveThread: false,
-          });
-        }
-
-        const config = memory.getMergedThreadConfig(memoryConfig || {});
-        const hasResourceScopeSemanticRecall =
-          (typeof config?.semanticRecall === 'object' && config?.semanticRecall?.scope !== 'thread') ||
-          config?.semanticRecall === true;
-        let [memoryResult, memorySystemMessage] = await Promise.all([
-          existingThread || hasResourceScopeSemanticRecall
-            ? this.getMemoryMessages({
-                resourceId,
-                threadId: threadObject.id,
-                vectorMessageSearch: new MessageList().add(messages, `user`).getLatestUserContent() || '',
-                memoryConfig,
-                requestContext,
-              })
-            : { messages: [] },
-          memory.getSystemMessage({ threadId: threadObject.id, resourceId, memoryConfig }),
-        ]);
-
-        const memoryMessages = memoryResult.messages;
-
-        this.logger.debug('Fetched messages from memory', {
-          threadId: threadObject.id,
-          runId,
-          fetchedCount: memoryMessages.length,
-        });
-
-        // So the agent doesn't get confused and start replying directly to messages
-        // that were added via semanticRecall from a different conversation,
-        // we need to pull those out and add to the system message.
-        const resultsFromOtherThreads = memoryMessages.filter(m => m.threadId !== threadObject.id);
-        if (resultsFromOtherThreads.length && !memorySystemMessage) {
-          memorySystemMessage = ``;
-        }
-        if (resultsFromOtherThreads.length) {
-          memorySystemMessage += `\nThe following messages were remembered from a different conversation:\n<remembered_from_other_conversation>\n${(() => {
-            let result = ``;
-
-            const messages = new MessageList().add(resultsFromOtherThreads, 'memory').get.all.v1();
-            let lastYmd: string | null = null;
-            for (const msg of messages) {
-              const date = msg.createdAt;
-              const year = date.getUTCFullYear();
-              const month = date.toLocaleString('default', { month: 'short' });
-              const day = date.getUTCDate();
-              const ymd = `${year}, ${month}, ${day}`;
-              const utcHour = date.getUTCHours();
-              const utcMinute = date.getUTCMinutes();
-              const hour12 = utcHour % 12 || 12;
-              const ampm = utcHour < 12 ? 'AM' : 'PM';
-              const timeofday = `${hour12}:${utcMinute < 10 ? '0' : ''}${utcMinute} ${ampm}`;
-
-              if (!lastYmd || lastYmd !== ymd) {
-                result += `\nthe following messages are from ${ymd}\n`;
-              }
-              result += `
-  Message ${msg.threadId && msg.threadId !== threadObject.id ? 'from previous conversation' : ''} at ${timeofday}: ${JSON.stringify(msg)}`;
-
-              lastYmd = ymd;
-            }
-            return result;
-          })()}\n<end_remembered_from_other_conversation>`;
-        }
-
-        if (memorySystemMessage) {
-          messageList.addSystem(memorySystemMessage, 'memory');
-        }
-
-        messageList
-          .add(
-            memoryMessages.filter((m: MastraDBMessage) => m.threadId === threadObject.id), // filter out messages from other threads. those are added to system message above
-            'memory',
-          )
-          // add new user messages to the list AFTER remembered messages to make ordering more reliable
-          .add(messages, 'user');
-
-        const { tripwireTriggered, tripwireReason } = await this.__runInputProcessors({
-          requestContext,
-          tracingContext: innerTracingContext,
-          messageList,
-        });
-
-        const systemMessages = messageList.getSystemMessages();
-
-        const systemMessage =
-          [...systemMessages, ...messageList.getSystemMessages('memory')]?.map(m => m.content)?.join(`\n`) ?? undefined;
-
-        const processedMemoryMessages = await memory.processMessages({
-          // these will be processed
-          messages: messageList.get.remembered.v1() as CoreMessage[],
-          // these are here for inspecting but shouldn't be returned by the processor
-          // - ex TokenLimiter needs to measure all tokens even though it's only processing remembered messages
-          newMessages: messageList.get.input.v1() as CoreMessage[],
-          systemMessage,
-          memorySystemMessage: memorySystemMessage || undefined,
-        });
-
-        const processedList = new MessageList({
-          threadId: threadObject.id,
-          resourceId,
-          generateMessageId: this.#mastra?.generateId?.bind(this.#mastra),
-          // @ts-ignore Flag for agent network messages
-          _agentNetworkAppend: this._agentNetworkAppend,
-        })
-          .addSystem(instructions || (await this.getInstructions({ requestContext })))
-          .addSystem(memorySystemMessage)
-          .addSystem(systemMessages)
-          .add(context || [], 'context')
-          .add(processedMemoryMessages, 'memory')
-          .add(messageList.get.input.db(), 'user')
-          .get.all.prompt();
-
-        return {
-          convertedTools,
-          thread: threadObject,
-          messageList,
-          // add old processed messages + new input messages
-          messageObjects: processedList,
-          agentAISpan,
-          ...(tripwireTriggered && {
-            tripwire: true,
-            tripwireReason,
-          }),
-          threadExists: !!existingThread,
-        };
-      },
-      after: async ({
-        result,
-        thread: threadAfter,
-        threadId,
-        memoryConfig,
-        outputText,
-        runId,
-        messageList,
-        threadExists,
-        structuredOutput = false,
-        overrideScorers,
-        agentAISpan,
-      }: {
-        runId: string;
-        result: Record<string, any>;
-        thread: StorageThreadType | null | undefined;
-        threadId?: string;
-        memoryConfig: MemoryConfig | undefined;
-        outputText: string;
-        messageList: MessageList;
-        threadExists: boolean;
-        structuredOutput?: boolean;
-        overrideScorers?: MastraScorers;
-        agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-      }) => {
-        const resToLog = {
-          text: result?.text,
-          object: result?.object,
-          toolResults: result?.toolResults,
-          toolCalls: result?.toolCalls,
-          usage: result?.usage,
-          steps: result?.steps?.map((s: any) => {
-            return {
-              stepType: s?.stepType,
-              text: result?.text,
-              object: result?.object,
-              toolResults: result?.toolResults,
-              toolCalls: result?.toolCalls,
-              usage: result?.usage,
-            };
-          }),
-        };
-
-        this.logger.debug(`[Agent:${this.name}] - Post processing LLM response`, {
-          runId,
-          result: resToLog,
-          threadId,
-        });
-
-        const messageListResponses = new MessageList({
-          threadId,
-          resourceId,
-          generateMessageId: this.#mastra?.generateId?.bind(this.#mastra),
-          // @ts-ignore Flag for agent network messages
-          _agentNetworkAppend: this._agentNetworkAppend,
-        })
-          .add(result.response.messages, 'response')
-          .get.all.core();
-
-        const usedWorkingMemory = messageListResponses?.some(
-          m => m.role === 'tool' && m?.content?.some(c => c?.toolName === 'updateWorkingMemory'),
-        );
-        // working memory updates the thread, so we need to get the latest thread if we used it
-        const memory = await this.getMemory({ requestContext });
-        const thread = usedWorkingMemory
-          ? threadId
-            ? await memory?.getThreadById({ threadId })
-            : undefined
-          : threadAfter;
-
-        if (memory && resourceId && thread) {
-          try {
-            // Add LLM response messages to the list
-            let responseMessages = result.response.messages;
-            if (!responseMessages && result.object) {
-              responseMessages = [
-                {
-                  role: 'assistant',
-                  content: [
-                    {
-                      type: 'text',
-                      text: outputText, // outputText contains the stringified object
-                    },
-                  ],
-                },
-              ];
-            }
-            if (responseMessages) {
-              messageList.add(responseMessages, 'response');
-            }
-
-            if (!threadExists) {
-              await memory.createThread({
-                threadId: thread.id,
-                metadata: thread.metadata,
-                title: thread.title,
-                memoryConfig,
-                resourceId: thread.resourceId,
-              });
-            }
-
-            // Parallelize title generation and message saving
-            const promises: Promise<any>[] = [saveQueueManager.flushMessages(messageList, threadId, memoryConfig)];
-
-            // Add title generation to promises if needed
-            if (thread.title?.startsWith('New Thread')) {
-              const config = memory.getMergedThreadConfig(memoryConfig);
-              const userMessage = this.getMostRecentUserMessage(messageList.get.all.ui());
-
-              const {
-                shouldGenerate,
-                model: titleModel,
-                instructions: titleInstructions,
-              } = this.resolveTitleGenerationConfig(config?.generateTitle);
-
-              if (shouldGenerate && userMessage) {
-                promises.push(
-                  this.genTitle(
-                    userMessage,
-                    requestContext,
-                    { currentSpan: agentAISpan },
-                    titleModel,
-                    titleInstructions,
-                  ).then(title => {
-                    if (title) {
-                      return memory.createThread({
-                        threadId: thread.id,
-                        resourceId,
-                        memoryConfig,
-                        title,
-                        metadata: thread.metadata,
-                      });
-                    }
-                  }),
-                );
-              }
-            }
-
-            await Promise.all(promises);
-          } catch (e) {
-            await saveQueueManager.flushMessages(messageList, threadId, memoryConfig);
-            if (e instanceof MastraError) {
-              agentAISpan?.error({ error: e });
-              throw e;
-            }
-            const mastraError = new MastraError(
-              {
-                id: 'AGENT_MEMORY_PERSIST_RESPONSE_MESSAGES_FAILED',
-                domain: ErrorDomain.AGENT,
-                category: ErrorCategory.SYSTEM,
-                details: {
-                  agentName: this.name,
-                  runId: runId || '',
-                  threadId: threadId || '',
-                  result: JSON.stringify(resToLog),
-                },
-              },
-              e,
-            );
-            this.logger.trackException(mastraError);
-            this.logger.error(mastraError.toString());
-            agentAISpan?.error({ error: mastraError });
-            throw mastraError;
-          }
-        } else {
-          let responseMessages = result.response.messages;
-          if (!responseMessages && result.object) {
-            responseMessages = [
-              {
-                role: 'assistant',
-                content: [
-                  {
-                    type: 'text',
-                    text: outputText, // outputText contains the stringified object
-                  },
-                ],
-              },
-            ];
-          }
-          if (responseMessages) {
-            messageList.add(responseMessages, 'response');
-          }
-        }
-
-        await this.#runScorers({
-          messageList,
-          runId,
-          requestContext,
-          structuredOutput,
-          overrideScorers,
-          threadId,
-          resourceId,
-          tracingContext: { currentSpan: agentAISpan },
-        });
-
-        const scoringData: {
-          input: Omit<ScorerRunInputForAgent, 'runId'>;
-          output: ScorerRunOutputForAgent;
-        } = {
-          input: {
-            inputMessages: messageList.getPersisted.input.ui(),
-            rememberedMessages: messageList.getPersisted.remembered.ui(),
-            systemMessages: messageList.getSystemMessages(),
-            taggedSystemMessages: messageList.getPersisted.taggedSystemMessages,
-          },
-          output: messageList.getPersisted.response.ui(),
-        };
-
-        agentAISpan?.end({
-          output: {
-            text: result?.text,
-            object: result?.object,
-            files: result?.files,
-          },
-        });
-
-        return {
-          scoringData,
-        };
-      },
-    };
-  }
-
   async #runScorers({
     messageList,
     runId,
@@ -2688,309 +2161,6 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
     }
 
     return result;
-  }
-
-  /**
-   * Prepares options and handlers for LLM text/object generation or streaming.
-   * @internal
-   */
-  private prepareLLMOptions<
-    Tools extends ToolSet,
-    Output extends ZodSchema | JSONSchema7 | undefined = undefined,
-    ExperimentalOutput extends ZodSchema | JSONSchema7 | undefined = undefined,
-  >(
-    messages: MessageListInput,
-    options: AgentGenerateOptions<Output, ExperimentalOutput>,
-    methodType: 'generate' | 'stream',
-  ): Promise<{
-    before: () => Promise<
-      Omit<
-        Output extends undefined
-          ? GenerateTextWithMessagesArgs<Tools, ExperimentalOutput>
-          : Omit<GenerateObjectWithMessagesArgs<NonNullable<Output>>, 'structuredOutput'> & {
-              output?: Output;
-              experimental_output?: never;
-            },
-        'runId'
-      > & { runId: string } & TripwireProperties & { agentAISpan?: AISpan<AISpanType.AGENT_RUN> }
-    >;
-    after: (args: {
-      result: GenerateReturn<any, Output, ExperimentalOutput>;
-      outputText: string;
-      structuredOutput?: boolean;
-      agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-      overrideScorers?:
-        | MastraScorers
-        | Record<string, { scorer: MastraScorer['name']; sampling?: ScoringSamplingConfig }>;
-    }) => Promise<{
-      scoringData: {
-        input: Omit<ScorerRunInputForAgent, 'runId'>;
-        output: ScorerRunOutputForAgent;
-      };
-    }>;
-    llm: MastraLLM;
-  }>;
-  /**
-   * @internal
-   */
-  private prepareLLMOptions<
-    Tools extends ToolSet,
-    Output extends ZodSchema | JSONSchema7 | undefined = undefined,
-    ExperimentalOutput extends ZodSchema | JSONSchema7 | undefined = undefined,
-  >(
-    messages: MessageListInput,
-    options: AgentStreamOptions<Output, ExperimentalOutput>,
-    methodType: 'generate' | 'stream',
-  ): Promise<{
-    before: () => Promise<
-      Omit<
-        Output extends undefined
-          ? StreamTextWithMessagesArgs<Tools, ExperimentalOutput>
-          : Omit<StreamObjectWithMessagesArgs<NonNullable<Output>>, 'structuredOutput'> & {
-              output?: Output;
-              experimental_output?: never;
-            },
-        'runId'
-      > & { runId: string } & TripwireProperties & { agentAISpan?: AISpan<AISpanType.AGENT_RUN> }
-    >;
-    after: (args: {
-      result: OriginalStreamTextOnFinishEventArg<any> | OriginalStreamObjectOnFinishEventArg<ExperimentalOutput>;
-      outputText: string;
-      structuredOutput?: boolean;
-      agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-      overrideScorers?:
-        | MastraScorers
-        | Record<string, { scorer: MastraScorer['name']; sampling?: ScoringSamplingConfig }>;
-    }) => Promise<{
-      scoringData: {
-        input: Omit<ScorerRunInputForAgent, 'runId'>;
-        output: ScorerRunOutputForAgent;
-      };
-    }>;
-    llm: MastraLLMV1;
-  }>;
-  /**
-   * @internal
-   */
-  private async prepareLLMOptions<
-    Tools extends ToolSet,
-    Output extends ZodSchema | JSONSchema7 | undefined = undefined,
-    ExperimentalOutput extends ZodSchema | JSONSchema7 | undefined = undefined,
-  >(
-    messages: MessageListInput,
-    options: (AgentGenerateOptions<Output, ExperimentalOutput> | AgentStreamOptions<Output, ExperimentalOutput>) & {
-      writableStream?: WritableStream<ChunkType>;
-    },
-    methodType: 'generate' | 'stream',
-  ): Promise<{
-    before:
-      | (() => Promise<
-          Omit<
-            Output extends undefined
-              ? StreamTextWithMessagesArgs<Tools, ExperimentalOutput>
-              : Omit<StreamObjectWithMessagesArgs<NonNullable<Output>>, 'structuredOutput'> & {
-                  output?: Output;
-                  experimental_output?: never;
-                },
-            'runId'
-          > & { runId: string } & TripwireProperties & { agentAISpan?: AISpan<AISpanType.AGENT_RUN> }
-        >)
-      | (() => Promise<
-          Omit<
-            Output extends undefined
-              ? GenerateTextWithMessagesArgs<Tools, ExperimentalOutput>
-              : Omit<GenerateObjectWithMessagesArgs<NonNullable<Output>>, 'structuredOutput'> & {
-                  output?: Output;
-                  experimental_output?: never;
-                },
-            'runId'
-          > & { runId: string } & TripwireProperties & { agentAISpan?: AISpan<AISpanType.AGENT_RUN> }
-        >);
-    after:
-      | ((args: {
-          result: GenerateReturn<any, Output, ExperimentalOutput>;
-          outputText: string;
-          agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-          overrideScorers?: MastraScorers;
-        }) => Promise<{
-          scoringData: {
-            input: Omit<ScorerRunInputForAgent, 'runId'>;
-            output: ScorerRunOutputForAgent;
-          };
-        }>)
-      | ((args: {
-          agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-          result: OriginalStreamTextOnFinishEventArg<any> | OriginalStreamObjectOnFinishEventArg<ExperimentalOutput>;
-          outputText: string;
-          structuredOutput?: boolean;
-          overrideScorers?: MastraScorers;
-        }) => Promise<{
-          scoringData: {
-            input: Omit<ScorerRunInputForAgent, 'runId'>;
-            output: ScorerRunOutputForAgent;
-          };
-        }>);
-    llm: MastraLLM;
-  }> {
-    const {
-      context,
-      memoryOptions: memoryConfigFromArgs,
-      resourceId: resourceIdFromArgs,
-      maxSteps,
-      onStepFinish,
-      toolsets,
-      clientTools,
-      temperature,
-      toolChoice = 'auto',
-      requestContext = new RequestContext(),
-      tracingContext,
-      tracingOptions,
-      savePerStep,
-      writableStream,
-      ...args
-    } = options;
-
-    // Currently not being used, but should be kept around for now in case it's needed later
-    // const generateMessageId =
-    //   `experimental_generateMessageId` in args && typeof args.experimental_generateMessageId === `function`
-    //     ? (args.experimental_generateMessageId as IDGenerator)
-    //     : undefined;
-
-    const threadFromArgs = resolveThreadIdFromArgs({ threadId: args.threadId, memory: args.memory });
-    const resourceId = args.memory?.resource || resourceIdFromArgs;
-    const memoryConfig = args.memory?.options || memoryConfigFromArgs;
-
-    if (resourceId && threadFromArgs && !this.hasOwnMemory()) {
-      this.logger.warn(
-        `[Agent:${this.name}] - No memory is configured but resourceId and threadId were passed in args. This will not work.`,
-      );
-    }
-    const runId = args.runId || this.#mastra?.generateId() || randomUUID();
-    const instructions = args.instructions || (await this.getInstructions({ requestContext }));
-    const llm = await this.getLLM({ requestContext });
-
-    const memory = await this.getMemory({ requestContext });
-    const saveQueueManager = new SaveQueueManager({
-      logger: this.logger,
-      memory,
-    });
-
-    const { before, after } = this.__primitive({
-      messages,
-      instructions,
-      context,
-      thread: threadFromArgs,
-      memoryConfig,
-      resourceId,
-      runId,
-      toolsets,
-      clientTools,
-      requestContext,
-      saveQueueManager,
-      writableStream,
-      methodType,
-      tracingContext,
-      tracingOptions,
-    });
-
-    let messageList: MessageList;
-    let thread: StorageThreadType | null | undefined;
-    let threadExists: boolean;
-
-    return {
-      llm,
-      before: async () => {
-        const beforeResult = await before();
-        const { messageObjects, convertedTools, agentAISpan } = beforeResult;
-        threadExists = beforeResult.threadExists || false;
-        messageList = beforeResult.messageList;
-        thread = beforeResult.thread;
-
-        const threadId = thread?.id;
-
-        // can't type this properly sadly :(
-        const result = {
-          ...options,
-          messages: messageObjects,
-          tools: convertedTools as Record<string, Tool>,
-          runId,
-          temperature,
-          toolChoice,
-          threadId,
-          resourceId,
-          requestContext,
-          onStepFinish: async (props: any) => {
-            if (savePerStep) {
-              if (!threadExists && memory && thread) {
-                await memory.createThread({
-                  threadId,
-                  title: thread.title,
-                  metadata: thread.metadata,
-                  resourceId: thread.resourceId,
-                  memoryConfig,
-                });
-                threadExists = true;
-              }
-
-              await this.saveStepMessages({
-                saveQueueManager,
-                result: props,
-                messageList,
-                threadId,
-                memoryConfig,
-                runId,
-              });
-            }
-
-            return onStepFinish?.({ ...props, runId });
-          },
-          ...(beforeResult.tripwire && {
-            tripwire: beforeResult.tripwire,
-            tripwireReason: beforeResult.tripwireReason,
-          }),
-          ...args,
-          agentAISpan,
-        } as any;
-
-        return result;
-      },
-      after: async ({
-        result,
-        outputText,
-        structuredOutput = false,
-        agentAISpan,
-        overrideScorers,
-      }:
-        | {
-            result: GenerateReturn<any, Output, ExperimentalOutput>;
-            outputText: string;
-            structuredOutput?: boolean;
-            agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-            overrideScorers?: MastraScorers;
-          }
-        | {
-            result: StreamReturn<any, Output, ExperimentalOutput>;
-            outputText: string;
-            structuredOutput?: boolean;
-            agentAISpan?: AISpan<AISpanType.AGENT_RUN>;
-            overrideScorers?: MastraScorers;
-          }) => {
-        const afterResult = await after({
-          result,
-          outputText,
-          threadId: thread?.id,
-          thread,
-          memoryConfig,
-          runId,
-          messageList,
-          structuredOutput,
-          threadExists,
-          agentAISpan,
-          overrideScorers,
-        });
-        return afterResult;
-      },
-    };
   }
 
   /**
@@ -3680,295 +2850,7 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
     messages: MessageListInput,
     generateOptions: AgentGenerateOptions<OUTPUT, EXPERIMENTAL_OUTPUT> = {},
   ): Promise<OUTPUT extends undefined ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT> : GenerateObjectResult<OUTPUT>> {
-    if ('structuredOutput' in generateOptions && generateOptions.structuredOutput) {
-      throw new MastraError({
-        id: 'AGENT_GENERATE_LEGACY_STRUCTURED_OUTPUT_NOT_SUPPORTED',
-        domain: ErrorDomain.AGENT,
-        category: ErrorCategory.USER,
-        text: 'This method does not support structured output. Please use generate() instead.',
-      });
-    }
-    const defaultGenerateOptionsLegacy = await this.getDefaultGenerateOptionsLegacy({
-      requestContext: generateOptions.requestContext,
-    });
-    const mergedGenerateOptions: AgentGenerateOptions<OUTPUT, EXPERIMENTAL_OUTPUT> = {
-      ...defaultGenerateOptionsLegacy,
-      ...generateOptions,
-      experimental_generateMessageId:
-        defaultGenerateOptionsLegacy.experimental_generateMessageId || this.#mastra?.generateId?.bind(this.#mastra),
-    };
-
-    const { llm, before, after } = await this.prepareLLMOptions(messages, mergedGenerateOptions, 'generate');
-
-    if (llm.getModel().specificationVersion !== 'v1') {
-      this.logger.error('V2 models are not supported for generateLegacy. Please use generate instead.', {
-        modelId: llm.getModel().modelId,
-      });
-
-      throw new MastraError({
-        id: 'AGENT_GENERATE_V2_MODEL_NOT_SUPPORTED',
-        domain: ErrorDomain.AGENT,
-        category: ErrorCategory.USER,
-        details: {
-          modelId: llm.getModel().modelId,
-        },
-        text: 'V2 models are not supported for generateLegacy. Please use generate instead.',
-      });
-    }
-
-    let llmToUse = llm as MastraLLMV1;
-
-    const beforeResult = await before();
-    const traceId = getValidTraceId(beforeResult.agentAISpan);
-
-    // Check for tripwire and return early if triggered
-    if (beforeResult.tripwire) {
-      const tripwireResult = {
-        text: '',
-        object: undefined,
-        usage: { totalTokens: 0, promptTokens: 0, completionTokens: 0 },
-        finishReason: 'other',
-        response: {
-          id: randomUUID(),
-          timestamp: new Date(),
-          modelId: 'tripwire',
-          messages: [],
-        },
-        responseMessages: [],
-        toolCalls: [],
-        toolResults: [],
-        warnings: undefined,
-        request: {
-          body: JSON.stringify({ messages: [] }),
-        },
-        experimental_output: undefined,
-        steps: undefined,
-        experimental_providerMetadata: undefined,
-        tripwire: true,
-        tripwireReason: beforeResult.tripwireReason,
-        traceId,
-      };
-
-      return tripwireResult as unknown as OUTPUT extends undefined
-        ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-        : GenerateObjectResult<OUTPUT>;
-    }
-
-    const { experimental_output, output, agentAISpan, ...llmOptions } = beforeResult;
-
-    const tracingContext: TracingContext = { currentSpan: agentAISpan };
-
-    // Handle structuredOutput option by creating an StructuredOutputProcessor
-    let finalOutputProcessors = mergedGenerateOptions.outputProcessors;
-
-    if (!output || experimental_output) {
-      const result = await llmToUse.__text<any, EXPERIMENTAL_OUTPUT>({
-        ...llmOptions,
-        tracingContext,
-        experimental_output,
-      });
-
-      const outputProcessorResult = await this.__runOutputProcessors({
-        requestContext: mergedGenerateOptions.requestContext || new RequestContext(),
-        tracingContext,
-        outputProcessorOverrides: finalOutputProcessors,
-        messageList: new MessageList({
-          threadId: llmOptions.threadId || '',
-          resourceId: llmOptions.resourceId || '',
-        }).add(
-          {
-            role: 'assistant',
-            content: [{ type: 'text', text: result.text }],
-          },
-          'response',
-        ),
-      });
-
-      // Handle tripwire for output processors
-      if (outputProcessorResult.tripwireTriggered) {
-        const tripwireResult = {
-          text: '',
-          object: undefined,
-          usage: { totalTokens: 0, promptTokens: 0, completionTokens: 0 },
-          finishReason: 'other',
-          response: {
-            id: randomUUID(),
-            timestamp: new Date(),
-            modelId: 'tripwire',
-            messages: [],
-          },
-          responseMessages: [],
-          toolCalls: [],
-          toolResults: [],
-          warnings: undefined,
-          request: {
-            body: JSON.stringify({ messages: [] }),
-          },
-          experimental_output: undefined,
-          steps: undefined,
-          experimental_providerMetadata: undefined,
-          tripwire: true,
-          tripwireReason: outputProcessorResult.tripwireReason,
-          traceId,
-        };
-
-        return tripwireResult as unknown as OUTPUT extends undefined
-          ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-          : GenerateObjectResult<OUTPUT>;
-      }
-
-      const newText = outputProcessorResult.messageList.get.response
-        .db()
-        .map(msg => msg.content.parts.map(part => (part.type === 'text' ? part.text : '')).join(''))
-        .join('');
-
-      // Update the result text with processed output
-      (result as any).text = newText;
-
-      // If there are output processors, check for structured data in message metadata
-      if (finalOutputProcessors && finalOutputProcessors.length > 0) {
-        // First check if any output processor provided structured data via metadata
-        const messages = outputProcessorResult.messageList.get.response.db();
-        this.logger.debug(
-          'Checking messages for experimentalOutput metadata:',
-          messages.map(m => ({
-            role: m.role,
-            hasContentMetadata: !!m.content.metadata,
-            contentMetadata: m.content.metadata,
-          })),
-        );
-
-        const messagesWithStructuredData = messages.filter(
-          msg => msg.content.metadata && msg.content.metadata.structuredOutput,
-        );
-
-        this.logger.debug('Messages with structured data:', messagesWithStructuredData.length);
-
-        if (messagesWithStructuredData[0] && messagesWithStructuredData[0].content.metadata?.structuredOutput) {
-          // Use structured data from processor metadata for result.object
-          (result as any).object = messagesWithStructuredData[0].content.metadata.structuredOutput;
-          this.logger.debug('Using structured data from processor metadata for result.object');
-        } else {
-          // Fallback: try to parse text as JSON (original behavior)
-          try {
-            const processedOutput = JSON.parse(newText);
-            (result as any).object = processedOutput;
-            this.logger.debug('Using fallback JSON parsing for result.object');
-          } catch (error) {
-            this.logger.warn('Failed to parse processed output as JSON, updating text only', { error });
-          }
-        }
-      }
-
-      const overrideScorers = mergedGenerateOptions.scorers;
-      const afterResult = await after({
-        result: result as unknown as OUTPUT extends undefined
-          ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-          : GenerateObjectResult<OUTPUT>,
-        outputText: newText,
-        agentAISpan,
-        ...(overrideScorers ? { overrideScorers } : {}),
-      });
-
-      if (generateOptions.returnScorerData) {
-        result.scoringData = afterResult.scoringData;
-      }
-
-      result.traceId = traceId;
-
-      return result as unknown as OUTPUT extends undefined
-        ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-        : GenerateObjectResult<OUTPUT>;
-    }
-
-    const result = await llmToUse.__textObject<NonNullable<OUTPUT>>({
-      ...llmOptions,
-      tracingContext,
-      structuredOutput: output as NonNullable<OUTPUT>,
-    });
-
-    const outputText = JSON.stringify(result.object);
-
-    const outputProcessorResult = await this.__runOutputProcessors({
-      requestContext: mergedGenerateOptions.requestContext || new RequestContext(),
-      tracingContext,
-      messageList: new MessageList({
-        threadId: llmOptions.threadId || '',
-        resourceId: llmOptions.resourceId || '',
-      }).add(
-        {
-          role: 'assistant',
-          content: [{ type: 'text', text: outputText }],
-        },
-        'response',
-      ),
-    });
-
-    // Handle tripwire for output processors
-    if (outputProcessorResult.tripwireTriggered) {
-      const tripwireResult = {
-        text: '',
-        object: undefined,
-        usage: { totalTokens: 0, promptTokens: 0, completionTokens: 0 },
-        finishReason: 'other',
-        response: {
-          id: randomUUID(),
-          timestamp: new Date(),
-          modelId: 'tripwire',
-          messages: [],
-        },
-        responseMessages: [],
-        toolCalls: [],
-        toolResults: [],
-        warnings: undefined,
-        request: {
-          body: JSON.stringify({ messages: [] }),
-        },
-        experimental_output: undefined,
-        steps: undefined,
-        experimental_providerMetadata: undefined,
-        tripwire: true,
-        tripwireReason: outputProcessorResult.tripwireReason,
-        traceId,
-      };
-
-      return tripwireResult as unknown as OUTPUT extends undefined
-        ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-        : GenerateObjectResult<OUTPUT>;
-    }
-
-    const newText = outputProcessorResult.messageList.get.response
-      .db()
-      .map(msg => msg.content.parts.map(part => (part.type === 'text' ? part.text : '')).join(''))
-      .join('');
-
-    // Parse the processed text and update the result object
-    try {
-      const processedObject = JSON.parse(newText);
-      (result as any).object = processedObject;
-    } catch (error) {
-      this.logger.warn('Failed to parse processed output as JSON, keeping original result', { error });
-    }
-
-    const afterResult = await after({
-      result: result as unknown as OUTPUT extends undefined
-        ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-        : GenerateObjectResult<OUTPUT>,
-      outputText: newText,
-      ...(generateOptions.scorers ? { overrideScorers: generateOptions.scorers } : {}),
-      structuredOutput: true,
-      agentAISpan,
-    });
-
-    if (generateOptions.returnScorerData) {
-      result.scoringData = afterResult.scoringData;
-    }
-
-    result.traceId = traceId;
-
-    return result as unknown as OUTPUT extends undefined
-      ? GenerateTextResult<any, EXPERIMENTAL_OUTPUT>
-      : GenerateObjectResult<OUTPUT>;
+    return this.getLegacyHandler().generateLegacy(messages, generateOptions);
   }
 
   /**
@@ -4028,174 +2910,7 @@ export class Agent<TAgentId extends string = string, TTools extends ToolsInput =
     | StreamTextResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown>
     | (StreamObjectResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown, any> & TracingProperties)
   > {
-    const defaultStreamOptionsLegacy = await this.getDefaultStreamOptionsLegacy({
-      requestContext: streamOptions.requestContext,
-    });
-
-    const mergedStreamOptions: AgentStreamOptions<OUTPUT, EXPERIMENTAL_OUTPUT> = {
-      ...defaultStreamOptionsLegacy,
-      ...streamOptions,
-      experimental_generateMessageId:
-        defaultStreamOptionsLegacy.experimental_generateMessageId || this.#mastra?.generateId?.bind(this.#mastra),
-    };
-
-    const { llm, before, after } = await this.prepareLLMOptions(messages, mergedStreamOptions, 'stream');
-
-    if (llm.getModel().specificationVersion !== 'v1') {
-      this.logger.error('V2 models are not supported for streamLegacy. Please use stream instead.', {
-        modelId: llm.getModel().modelId,
-      });
-
-      throw new MastraError({
-        id: 'AGENT_STREAM_V2_MODEL_NOT_SUPPORTED',
-        domain: ErrorDomain.AGENT,
-        category: ErrorCategory.USER,
-        details: {
-          modelId: llm.getModel().modelId,
-        },
-        text: 'V2 models are not supported for streamLegacy. Please use stream instead.',
-      });
-    }
-
-    const beforeResult = await before();
-    const traceId = getValidTraceId(beforeResult.agentAISpan);
-
-    // Check for tripwire and return early if triggered
-    if (beforeResult.tripwire) {
-      // Return a promise that resolves immediately with empty result
-      const emptyResult = {
-        textStream: (async function* () {
-          // Empty async generator - yields nothing
-        })(),
-        fullStream: Promise.resolve('').then(() => {
-          const emptyStream = new (globalThis as any).ReadableStream({
-            start(controller: any) {
-              controller.close();
-            },
-          });
-          return emptyStream;
-        }),
-        text: Promise.resolve(''),
-        usage: Promise.resolve({ totalTokens: 0, promptTokens: 0, completionTokens: 0 }),
-        finishReason: Promise.resolve('other'),
-        tripwire: true,
-        tripwireReason: beforeResult.tripwireReason,
-        response: {
-          id: randomUUID(),
-          timestamp: new Date(),
-          modelId: 'tripwire',
-          messages: [],
-        },
-        toolCalls: Promise.resolve([]),
-        toolResults: Promise.resolve([]),
-        warnings: Promise.resolve(undefined),
-        request: {
-          body: JSON.stringify({ messages: [] }),
-        },
-        experimental_output: undefined,
-        steps: undefined,
-        experimental_providerMetadata: undefined,
-        traceId,
-        toAIStream: () =>
-          Promise.resolve('').then(() => {
-            const emptyStream = new (globalThis as any).ReadableStream({
-              start(controller: any) {
-                controller.close();
-              },
-            });
-            return emptyStream;
-          }),
-        get experimental_partialOutputStream() {
-          return (async function* () {
-            // Empty async generator for partial output stream
-          })();
-        },
-        pipeDataStreamToResponse: () => Promise.resolve(),
-        pipeTextStreamToResponse: () => Promise.resolve(),
-        toDataStreamResponse: () => new Response('', { status: 200, headers: { 'Content-Type': 'text/plain' } }),
-        toTextStreamResponse: () => new Response('', { status: 200, headers: { 'Content-Type': 'text/plain' } }),
-      };
-
-      return emptyResult as unknown as
-        | StreamTextResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown>
-        | (StreamObjectResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown, any> & TracingProperties);
-    }
-
-    const { onFinish, runId, output, experimental_output, agentAISpan, ...llmOptions } = beforeResult;
-
-    const overrideScorers = mergedStreamOptions.scorers;
-    const tracingContext: TracingContext = { currentSpan: agentAISpan };
-
-    if (!output || experimental_output) {
-      this.logger.debug(`Starting agent ${this.name} llm stream call`, {
-        runId,
-      });
-
-      const streamResult = llm.__stream({
-        ...llmOptions,
-        experimental_output,
-        tracingContext,
-        outputProcessors: await this.listResolvedOutputProcessors(mergedStreamOptions.requestContext),
-        onFinish: async result => {
-          try {
-            const outputText = result.text;
-            await after({
-              result,
-              outputText,
-              agentAISpan,
-              ...(overrideScorers ? { overrideScorers } : {}),
-            });
-          } catch (e) {
-            this.logger.error('Error saving memory on finish', {
-              error: e,
-              runId,
-            });
-          }
-          await onFinish?.({ ...result, runId } as any);
-        },
-        runId,
-      });
-
-      streamResult.traceId = traceId;
-
-      return streamResult as
-        | StreamTextResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown>
-        | (StreamObjectResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown, any> & TracingProperties);
-    }
-
-    this.logger.debug(`Starting agent ${this.name} llm streamObject call`, {
-      runId,
-    });
-
-    const streamObjectResult = llm.__streamObject({
-      ...llmOptions,
-      tracingContext,
-      onFinish: async result => {
-        try {
-          const outputText = JSON.stringify(result.object);
-          await after({
-            result,
-            outputText,
-            structuredOutput: true,
-            agentAISpan,
-            ...(overrideScorers ? { overrideScorers } : {}),
-          });
-        } catch (e) {
-          this.logger.error('Error saving memory on finish', {
-            error: e,
-            runId,
-          });
-        }
-        await onFinish?.({ ...result, runId } as any);
-      },
-      runId,
-      structuredOutput: output,
-    });
-
-    (streamObjectResult as any).traceId = traceId;
-
-    return streamObjectResult as StreamObjectResult<any, OUTPUT extends ZodSchema ? z.infer<OUTPUT> : unknown, any> &
-      TracingProperties;
+    return this.getLegacyHandler().streamLegacy(messages, streamOptions);
   }
 
   /**
