+++ packages/core/src/agent/utils.ts
@@ -1,4 +1,5 @@
 import { ErrorCategory, ErrorDomain, MastraError } from '../error';
+import type { StorageThreadType } from '../memory';
 import type { OutputSchema } from '../stream';
 import type { Agent } from './agent';
 import type { AgentExecutionOptions } from './agent.types';
@@ -60,4 +61,17 @@ export async function tryStreamWithJsonFallback<
       structuredOutput: { ...options.structuredOutput, jsonPromptInjection: true },
     });
   }
+}
+
+export function resolveThreadIdFromArgs(args: {
+  memory?: { thread?: string | { id: string } };
+  threadId?: string;
+}): (Partial<StorageThreadType> & { id: string }) | undefined {
+  if (args?.memory?.thread) {
+    if (typeof args.memory.thread === 'string') return { id: args.memory.thread };
+    if (typeof args.memory.thread === 'object' && args.memory.thread.id)
+      return args.memory.thread as Partial<StorageThreadType> & { id: string };
+  }
+  if (args?.threadId) return { id: args.threadId };
+  return undefined;
 }
